fixed bugs

+++ socnavenv/envs/utils/object.py
@@ -34,22 +34,22 @@ class Object(object):
         self.one_hot_encoding = None
 
         if self.name == "robot":
-            self.one_hot_encoding = np.array([1, 0, 0, 0, 0, 0])
+            self.one_hot_encoding = np.array([1, 0, 0, 0, 0, 0], dtype=np.float32)
 
         elif self.name == "human":
-            self.one_hot_encoding = np.array([0, 1, 0, 0, 0, 0])
+            self.one_hot_encoding = np.array([0, 1, 0, 0, 0, 0], dtype=np.float32)
         
         elif self.name == "table":
-            self.one_hot_encoding = np.array([0, 0, 1, 0, 0, 0])
+            self.one_hot_encoding = np.array([0, 0, 1, 0, 0, 0], dtype=np.float32)
         
         elif self.name == "laptop":
-            self.one_hot_encoding = np.array([0, 0, 0, 1, 0, 0])
+            self.one_hot_encoding = np.array([0, 0, 0, 1, 0, 0], dtype=np.float32)
         
         elif self.name == "plant":
-            self.one_hot_encoding = np.array([0, 0, 0, 0, 1, 0])
+            self.one_hot_encoding = np.array([0, 0, 0, 0, 1, 0], dtype=np.float32)
         
         elif self.name == "wall":
-            self.one_hot_encoding = np.array([0, 0, 0, 0, 0, 1])
+            self.one_hot_encoding = np.array([0, 0, 0, 0, 0, 1], dtype=np.float32)
 
     def set(self, x, y, theta):
         """
